requirements.txt
Flask==2.3.3
azure-cognitiveservices-vision-computervision==0.9.0
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
Werkzeug==2.3.7
msrest==0.7.1
azure-common==1.1.28
deep-translator==1.11.4
opencensus-ext-azure==1.1.9
opencensus-ext-flask==0.8.2
opencensus-ext-logging==0.1.0